move latex funct to utils
